make LRectangle.getGeometry return correct type
LDrawRectangle returns a LPolygon. getGeometry for that then returns a JTS Rectangle. 
It would not make sense to return a linearRing here and thus have incompatible behavior.

Example use-case:  you create a LDrawRectangle from a custom menu button. The user draws the object, you save that as Geometry somewhere. The user opens contextmenu for that geometry and selects "Edit". User finishes editing and now the geometry which previously was a Polygon suddenly is a LinearLing.